requirements.txt
selenium==4.21.0
pandas==2.2.2
google-auth==2.29.0 
webdriver-manager==4.0.1 
python-dotenv==1.0.1